Removed custom tuple_index_of template struct.
The structure was used to lookup specific data types in std::tuples by index. When the same type of data is included in the tuple it only would return the first type. A std::tuple is able to return data for a specific type. It is no scenario that a std::tuple in this system will contain the same type twice. Hence std::get<T>() for tuple is enough.